refactor(frontend): handle structured sync errors
